Updated DigitalMinerUsage
Right now it's set to 100J/t which means it can be run with 4 heat generators (if each is set up optimal they'll produce 25 J/t). The Digital Miner is endgame and mines specific ores at a decent rate. With the current change it would use 12 instead of 4 (if it runs without speed upgrades and silk touch which is still decently fast.) making it harder to reach endgame status / unlimited ore. The ultimate number needs tweaking.